コードの整理・書き直し
