+++ Cargo.toml
@@ -1,29 +1,16 @@
 [package]
 name = "tigtag"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["hyoi"]
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-# [features]
-# default = [
-#   "bevy/bevy_gltf",
-#   "bevy/bevy_winit",
-#   "bevy/render",
-#   "bevy/png",
-# ]
-# native = [ "bevy/bevy_wgpu" ]
-# web    = [ "bevy_webgl2"    ]
 
 [dependencies]
 bevy = "0.6.0"
 bevy_prototype_lyon = "0.4.0"
 rand = "0.8.4"
 
-# bevy = { version = "0.5.0", default-features = false }
-# bevy_webgl2 = { version = "0.5.2", optional = true }
-# bevy_prototype_lyon = "0.3.1"
-
-# winit = "0.24.0"
-# getrandom = { version = "0.2.3", features = [ "wasm-bindgen" ] }
+# WASMの場合にどれか指定する必要がある？？
+#
+# rand = { version = "0.8.4", features = [ "wasm-bindgen" ] }
+# getrandom = { version = "0.2.4", features = [ "js" ] }
+# getrandom = { version = "0.2.4", features = [ "wasm-bindgen" ] }
