--- src/consts/mod.rs
@@ -1,21 +1,0 @@
-use super::*;
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-//定義と定数
-
-//マップの縦横のマス数
-pub const MAP_WIDTH : usize = 31; //43, 31, 25
-pub const MAP_HEIGHT: usize = 21; //21, 21, 17
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-//Sub module
-mod text_ui;
-
-//Re export
-pub use text_ui::*;
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-//End of code.
